no more double[] in the regular series as it is incorrect !
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1484 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-solver/choco-kernel/src/main/java/choco/kernel/solver/constraints/global/automata/fast_multicostregular/structure/StoredDirectedMultiGraph.java
@@ -104,7 +104,7 @@ public class StoredDirectedMultiGraph {
 
 
 
-    public StoredDirectedMultiGraph(IEnvironment environment, AbstractIntSConstraint constraint, DirectedMultigraph<Node, Arc> graph, int[][] layers, int[] starts, int[] offsets, int supportLength,double[][][][] costs)
+    public StoredDirectedMultiGraph(IEnvironment environment, AbstractIntSConstraint constraint, DirectedMultigraph<Node, Arc> graph, int[][] layers, int[] starts, int[] offsets, int supportLength,int[][][][] costs)
     {
         this.nbR = costs[0][0].length;
         this.constraint = constraint;
